requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_gnu
apispec=5.1.1=pypi_0
arm_pyart=1.11.6=py38h6c62de6_1
attrs=21.4.0=pypi_0
blas=1.0=mkl
boto3=1.20.46=pypi_0
botocore=1.23.46=pypi_0
bottleneck=1.3.2=py38heb32a55_1
brotli=1.0.9=he6710b0_2
bzip2=1.0.8=h7b6447c_0
c-ares=1.18.1=h7f8727e_0
ca-certificates=2021.10.26=h06a4308_2
certifi=2021.10.8=py38h06a4308_2
cftime=1.5.1.1=py38hce1f21e_0
click=8.0.3=pyhd3eb1b0_0
curl=7.80.0=h7f8727e_0
cycler=0.11.0=pyhd3eb1b0_0
dataclasses=0.8=pyh6d0b6a4_7
flasgger=0.9.5=pypi_0
flask=2.0.2=pyhd3eb1b0_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=h70c0345_0
giflib=5.2.1=h7b6447c_0
hdf4=4.2.13=h3ca952b_2
hdf5=1.10.6=hb1b8bf9_0
importlib-resources=5.4.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
itsdangerous=2.0.1=pyhd3eb1b0_0
jinja2=3.0.2=pyhd3eb1b0_0
jmespath=0.10.0=pypi_0
jpeg=9d=h7f8727e_0
jsonschema=4.4.0=pypi_0
kiwisolver=1.3.1=py38h2531618_0
krb5=1.19.2=hac12032_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libcurl=7.80.0=h0b77cf5_0
libedit=3.1.20210910=h7f8727e_0
libev=4.33=h7f8727e_1
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1d223b6_12
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=11.2.0=h1d223b6_12
libnetcdf=4.8.1=h42ceab0_1
libnghttp2=1.46.0=hce63b2e_0
libpng=1.6.37=hbc83047_0
libssh2=1.9.0=h1ba5d50_1
libstdcxx-ng=9.3.0=hd4cf53a_17
libtiff=4.2.0=h85742a9_0
libwebp=1.2.0=h89dd481_0
libwebp-base=1.2.0=h27cfd23_0
libzip=1.5.1=h8d318fa_1002
lz4-c=1.9.3=h295c915_1
markupsafe=2.0.1=py38h27cfd23_0
matplotlib-base=3.5.0=py38h3ed280b_0
mistune=2.0.2=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
munkres=1.1.4=py_0
ncurses=6.3=h7f8727e_2
netcdf4=1.5.7=py38ha0f2276_1
nexradaws=1.1=pypi_0
numexpr=2.8.1=py38h6abb31d_0
numpy=1.21.2=py38h20f2e39_0
numpy-base=1.21.2=py38h79a1101_0
olefile=0.46=pyhd3eb1b0_0
openssl=1.1.1m=h7f8727e_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.3.5=py38h8c16a72_0
pillow=8.4.0=py38h5aabda8_0
pip=21.2.4=py38h06a4308_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyrsistent=0.18.1=pypi_0
python=3.8.12=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.8=2_cp38
pytz=2021.3=pyhd3eb1b0_0
pyyaml=6.0=pypi_0
readline=8.1.2=h7f8727e_1
s3transfer=0.5.0=pypi_0
scipy=1.7.3=py38hc147768_0
setuptools=58.0.4=py38h06a4308_0
six=1.16.0=pyhd3eb1b0_0
sqlite=3.37.0=hc218d9a_0
tk=8.6.11=h1ccaba5_0
trmm_rsl=1.49=3
typing-extensions=3.10.0.2=hd3eb1b0_0
typing_extensions=3.10.0.2=pyh06a4308_0
urllib3=1.26.8=pypi_0
werkzeug=2.0.2=pyhd3eb1b0_0
wheel=0.37.1=pyhd3eb1b0_0
xarray=0.20.1=pyhd3eb1b0_1
xz=5.2.5=h7b6447c_0
zipp=3.7.0=pyhd3eb1b0_0
zlib=1.2.11=h7f8727e_4
zstd=1.4.9=haebb681_0
